+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+
+- convergent validation creators `validate` and `remains`
+
 ## [0.9.1] - 2018-10-23
 
 ### Changed